pyproject.toml
[tool.poetry]
name = "quickdna"
version = "0.1.0"
description = "Fast implementation of DNA translation in Rust"
authors = ["Theia Vogel <theia@vgel.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^6.1"
pylint = "^2.6"
flake8 = "^3.9"
wheel = "*"
pytest-runner = "*"
pytest-benchmark = "*"
hypothesis = "*"
black = "*"
maturin = "==0.9.4"
biopython = "^1.79"
